Improve project page main
