+++ frontend/src/javascripts/components/organisms/project-page-main.js
@@ -14,7 +14,7 @@ import {
   setDefaultProject,
 } from '../../actions/projects'
 
-import { PlanetImgs } from '../../constants/images'
+import FixedStar from '../atoms/fixed-star'
 import CircleOrbit from '../molecules/circle-orbit'
 import ProjectBar from '../molecules/project-bar'
 
@@ -42,74 +42,57 @@ class ProjectPage extends Component {
     }
   }
 
-  onDropPlanet(title, detail, deadline, planet_type, planet_size, orbit_pos) {
-    this.props.createAssignment(
-      title,
-      detail,
-      deadline,
-      planet_type,
-      planet_size,
-      orbit_pos,
-      this.props.match.params.projectId
-    )
-  }
+  // onDropPlanet(title, detail, deadline, planet_type, planet_size, orbit_pos) {
+  //   this.props.createAssignment(
+  //     title,
+  //     detail,
+  //     deadline,
+  //     planet_type,
+  //     planet_size,
+  //     orbit_pos,
+  //     this.props.match.params.projectId
+  //   )
+  // }
 
-  onDestroyPlanet(assignmentId) {
-    this.props.destroyAssignment(assignmentId)
-  }
+  // onDestroyPlanet(assignmentId) {
+  //   this.props.destroyAssignment(assignmentId)
+  // }
 
-  addPlanet(e) {
-    let target_class = e.target
-    let common_planet_tag = document.createElement('div')
-    let planet_tag = document.createElement('div')
-    let img_tag = document.createElement('img')
-    common_planet_tag.className =
-      'common top secundus-orbit-motion start-animation'
-    planet_tag.className = 'planet-medium-secundus'
-    img_tag.src = PlanetImgs.Uranus
+  // addPlanet(e) {
+  //   let target_class = e.target
+  //   let common_planet_tag = document.createElement('div')
+  //   let planet_tag = document.createElement('div')
+  //   let img_tag = document.createElement('img')
+  //   common_planet_tag.className =
+  //     'common top secundus-orbit-motion start-animation'
+  //   planet_tag.className = 'planet-medium-secundus'
+  //   img_tag.src = PlanetImgs.Uranus
 
-    planet_tag.appendChild(img_tag)
-    common_planet_tag.appendChild(planet_tag)
-    target_class.appendChild(common_planet_tag)
-  }
-
-  addSatelitePlanet() {
-    const conditional_class1 = document.getElementsByClassName(
-      'secundus-orbit-motion'
-    )
-    const conditional_class2 = document.getElementsByClassName(
-      'tertius-orbit-motion'
-    )
-    const target_class = document.getElementsByClassName('planet-large-primo')
-    const target_width = 1.5 * target_class[0].getBoundingClientRect().width
-
-    if (
-      conditional_class1[0].children[1].classList.contains('satelite-orbit')
-    ) {
-      conditional_class1[0].children[1].style.width = target_width + 'px'
-      conditional_class1[0].children[1].style.height = target_width + 'px'
-      conditional_class2[0].children[1].style.width = target_width + 'px'
-      conditional_class2[0].children[1].style.height = target_width + 'px'
-    }
-  }
+  //   planet_tag.appendChild(img_tag)
+  //   common_planet_tag.appendChild(planet_tag)
+  //   target_class.appendChild(common_planet_tag)
+  // }
 
   render() {
-    const { currentUser } = this.props
+    const {
+      currentUser,
+      match: { params },
+      currentProject,
+    } = this.props
+
     if (!currentUser) return <div>Loading....</div>
 
-    if (currentUser.id != this.props.match.params.userId) {
+    if (currentUser.id != params.userId) {
       const correctPath = `/users/${currentUser.id}`
       return <Redirect to={correctPath} />
     }
 
-    if (!this.props.currentProject) return <div>Loading....</div>
+    if (!currentProject) return <div>Loading....</div>
 
     return (
       <div>
         <div id="project-orbit">
-          <div id="fixed-star" onClick={this.addSatelitePlanet.bind(this)}>
-            <img src={PlanetImgs[this.props.currentProject.fixed_star_type]} />
-          </div>
+          <FixedStar fixedStarType={currentProject.fixed_star_type} />
           <CircleOrbit orbit="primo" />
           <CircleOrbit orbit="secundus" />
           <CircleOrbit orbit="tertius" />
